pyproject.toml
[project]
name = "chat-app"
version = "0.1.0"
description = "Мессенджер с возможностью обмена сообщениями и создания групповых чатов, аналогичный Telegram."
authors = [
    {name = "Степаненко Игорь Владимирович", email = "ighor_stepanenko@list.ru"},
]
dependencies = [
    "fastapi>=0.95.2",
    "uvicorn>=0.22.0",
    "python-jose>=3.3.0",
    "passlib>=1.7.4",
    "sqlalchemy>=2.0.15",
    "asyncpg>=0.27.0",
    "alembic>=1.11.1",
    "python-dotenv>=1.0.0",
    "websockets>=11.0.3",
    "pydantic>=1.10.7",
    "pydantic-settings>=2.0.3",
    "email-validator>=2.2.0",
    "python-multipart>=0.0.20",
]
requires-python = "==3.12.*"
readme = "README.md"

[tool.pdm]
distribution = false

[dependency-groups]
dev = [
    "mypy==1.11.2",
    "ruff==0.6.4",
]


[tool.ruff]
line-length = 120
indent-width = 4
target-version = "py312"

[tool.ruff.lint]
select = [
    "F" ,  # Pyflakes
    "E",  # pycodestyle. Error
    "W",  # pycodestyle. Warning
    "C90",  # mccabe
    "I",  # isort
    "N",  # pep8-naming
    "D",  # pydocstyle
    "UP",  # pyupgrade
    "YTT",  # flake8-2020
    "ASYNC",  # flake8-async
    "TRIO",  # flake8-trio
    "S",  # flake8-bandit
    "BLE",  # flake8-blind-except
    "FBT",  # flake8-boolean-trap
    "B",  # flake8-bugbear
    "A",  # flake8-builtins
    "COM",  # flake8-commas
    "C4",  # flake8-comprehensions
    "DTZ",  # flake8-datetimez
    "T10",  # flake8-debugger
    "DJ",  # flake8-django
    "EM",  # flake8-errmsg
    "EXE",  # flake8-executable
    "ISC",  # flake8-implicit-str-concat
    "ICN",  # flake8-import-conventions
    "LOG",  # flake8-logging
    "G",  # flake8-logging-format
    "INP",  # flake8-no-pep420
    "PIE",  # flake8-pie
    "T20",  # flake8-print
    "PYI",  # flake8-pyi
    "PT",  # flake8-pytest-style
    "Q",  # flake8-quotes
    "RSE",  # flake8-raise
    "RET",  # flake8-return
    "SLF",  # flake8-self
    "SLOT",  # flake8-slots
    "SIM",  # flake8-simplify
    "TID",  # flake8-tidy-imports
    "TCH",  # flake8-type-checking
    "INT",  # flake8-gettext
    "ARG",  # flake8-unused-arguments
    "PTH",  # flake8-use-pathlib
    "TD",   # flake8-todos
    "FIX",  # flake8-fixme
    "ERA",  # eradicate
    "PD",  # pandas-vet
    "PGH",  # pygrep-hooks
    "PLC",  # Pylint. Convention
    "PLE",  # Pylint. Error
    "PLR",  # Pylint. Refactor
    "PLW",  # Pylint. Warning
    "TRY",  # tryceratops
    "FLY",  # flynt
    "NPY",  # NumPy-specific rules
    "AIR",  # Airflow
    "PERF",  # Perflint
    "FURB",  # refurb
    "RUF",  # Ruff-specific rules
]
ignore = [
    "COM812",  # missing-trailing-comma: Trailing comma missing
    "D107",  # undocumented-public-init: Missing docstring in __init__
    "D203", # one-blank-line-before-class: 1 blank line required before class docstring
    "D205",  # blank-line-after-summary: 1 blank line required between summary line and description
    "D212",  # multi-line-summary-first-line: Multi-line docstring summary should start at the first line
    "RUF001",  # ambiguous-unicode-character-string: String contains ambiguous {}. Did you mean {}?
    "RUF002",  # ambiguous-unicode-character-docstring: Docstring contains ambiguous {}. Did you mean {}?
    "RUF003",  # ambiguous-unicode-character-comment: Comment contains ambiguous {}. Did you mean {}?
    "Q000",  # Single quotes found but double quotes preferred
    "B008",  # Function-call-in-argument-default: FastAPI Depends() intentionally used in arguments
]
fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.per-file-ignores]
"__init__.py" = [
    "D104",  # undocumented-public-package: Missing docstring in public package
]
"tests/*" = [
    "D103", # Missing docstring in public function
    "S101",  # assert: Use of `assert` detected
    "PLR2004",  # Magic value used in comparison
    "DTZ001",   # datetime.datetime()` called without a `tzinfo` argument
]


[tool.ruff.format]
quote-style = "single"
indent-style = "space"
skip-magic-trailing-comma = true
line-ending = "auto"
docstring-code-format = true
docstring-code-line-length = "dynamic"


[tool.mypy]
allow_subclassing_any = true
ignore_missing_imports = true
implicit_reexport = true
strict = true
disallow_untyped_decorators = false

